fix: 랭킹  신뢰도 정확도 오류 수정

See merge request s11-fintech-finance-sub2/S11P22B201!183

+++ backend/stockolm/src/main/java/com/example/stockolm/domain/stock/dto/response/BestAnalystResponse.java
@@ -13,9 +13,9 @@ public class BestAnalystResponse {
 
     private String userImagePath;
 
-    private Double reliability;
+    private Long reliability;
 
-    private Double accuracy;
+    private Long accuracy;
 
     private Long analystBoardId;
 
@@ -26,7 +26,9 @@ public class BestAnalystResponse {
     private Integer goalStock;
 
 
-    public BestAnalystResponse(Long analystBoardId, LocalDate goalDate, String opinion, Integer goalStock, Double reliability, Double accuracy, String analystName, String userImagePath, String analystNickname) {
+    public BestAnalystResponse(Long analystBoardId, LocalDate goalDate, String opinion,
+                               Integer goalStock, Long reliability, Long accuracy,
+                               String analystName, String userImagePath, String analystNickname) {
         this.analystBoardId = analystBoardId;
         this.goalDate = goalDate;
         this.opinion = opinion;
